Discard message when the archiver is run
This may be fixed in the chaton client protocol, but currently,
when the archiver truncates the current data file, chaton viewer
sends the entire content of the turncated file, which contains
old posts.  The case can be detected by comparing pos in the message
with the last read pos stored in the client.

Cf. https://github.com/shirok/Chaton/issues/4